Move SchemaSqlReadService to Virtual Node Write Service
