+++ components/virtual-node/virtual-node-write-service-impl/build.gradle
@@ -25,6 +25,7 @@ dependencies {
     implementation project(':components:membership:membership-group-read')
     implementation project(':components:membership:membership-client')
     implementation project(':components:membership:membership-persistence-client')
+    implementation project(':components:virtual-node:virtual-node-info-read-service')
     implementation project(':libs:configuration:configuration-core')
     implementation project(':libs:configuration:configuration-datamodel')
     implementation project(':libs:db:db-admin')
@@ -33,6 +34,7 @@ dependencies {
     implementation project(':libs:db:db-orm')
     implementation project(':libs:layered-property-map')
     implementation project(':libs:lifecycle:lifecycle')
+    implementation project(':libs:rest:rest')
     implementation project(':libs:serialization:serialization-avro')
     implementation project(':libs:virtual-node:virtual-node-common')
     implementation project(':libs:virtual-node:virtual-node-datamodel')
